# auth/app/swagger/model_create_map_key_val.go
/*
 * Occupant API
 *
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * API version: 0.0.1
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */
package swagger

type CreateMapKeyVal struct {
	Val string `json:"val"`
}