# auth/app/swagger/model_create_user.go
/*
 * Occupant API
 *
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * API version: 0.0.1
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */
package swagger

type CreateUser struct {
	Username string `json:"username,omitempty"`
	Password string `json:"password,omitempty"`
	Email string `json:"email,omitempty"`
}