# auth/app/swagger/model_token.go
/*
 * Occupant API
 *
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * API version: 0.0.1
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */
package swagger

type Token struct {
	Token string `json:"token,omitempty"`
	RefreshToken string `json:"refresh_token,omitempty"`
}