# auth/app/swagger/swagger.yml
openapi: 3.0.0
info:
  title: Occupant API
  version: 0.0.1

tags:
  - name: Occupant
    description: "Occupant - main entity"
  - name: Authorization
components:
  headers:
    RequestSuccess:
      schema:
        type: string
      description: "if the application has successfully processed the request, it returns success. If an error occurs during the processing of the request, it returns fail."

  responses:
    minimalResponse:
      description: "OK"

    NotFound:
      description: Not Found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/APIError"

    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/APIError"

    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/APIError"

    Conflict:
      description: Resource Already Exist
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/APIError"

    ImaTeapot:
      description: Code Execution Error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/APIError"

    InternalError:
      description: Internal Server Error

  schemas:
    APIError:
      type: object
      properties:
        error_code:
          type: string
        developer_message:
          type: string
      required:
        - error_code
        - message
        - developer_message

    Token:
      type: object
      properties:
        token:
          type: string
        refresh_token:
          type: string

    User:
      type: object
      properties:
        username:
          type: string
        password:
          type: string

    CreateUser:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
        email:
          type: string
    MapKeyVal:
      type: object
      properties:
        val:
          type: string

    CreateMapKeyVal:
      allOf:
        - $ref: "#/components/schemas/MapKeyVal"
        - required:
            - val
paths:
  /api/auth:
    post:
      summary: "auth by username and password "
      tags:
        - Authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
      description: ""
      responses:
        '200':
          description: "OK"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Token"
        '401':
          $ref: "#/components/responses/Unauthorized"
        '400':
          $ref: "#/components/responses/BadRequest"
        '418':
          $ref: "#/components/responses/ImaTeapot"
        '500':
          $ref: "#/components/responses/InternalError"
    put:
      summary: "get access using refresh token"
      tags:
        - Authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/Token"
                - required:
                    - refresh_token
      description: "Авторизация пользователя, обмен refresh_token на access_token для доступа к методам API"
      responses:
        '201':
          description: "OK"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Token"
        '401':
          $ref: "#/components/responses/Unauthorized"
        '400':
          $ref: "#/components/responses/BadRequest"
        '418':
          $ref: "#/components/responses/ImaTeapot"
        '500':
          $ref: "#/components/responses/InternalError"
  /api/signup:
    post:
      summary: "create new user"
      tags:
        - Authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/CreateUser"
      description: "Авторизация пользователя, получение токена доступа к методам API"
      responses:
        '201':
          description: "OK"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Token"
        '401':
          $ref: "#/components/responses/Unauthorized"
        '400':
          $ref: "#/components/responses/BadRequest"
        '418':
          $ref: "#/components/responses/ImaTeapot"
        '500':
          $ref: "#/components/responses/InternalError"
  /api/mapdb/{key}:
    get:
      summary: "get key"
      parameters:
        - in: path
          name: key
          required: true
          schema:
            type: string
      tags:
        - Mapdb
      description: "Get key"
      responses:
        '200':
          description: "OK"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MapKeyVal"
        '400':
          $ref: "#/components/responses/BadRequest"
        '418':
          $ref: "#/components/responses/ImaTeapot"
        '500':
          $ref: "#/components/responses/InternalError"
    post:
      summary: "create mapdb key"
      tags:
        - Mapdb
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMapKeyVal"
      description: "create mapdb key"
      responses:
        '201':
          headers:
            Location:
              schema:
                type: string
              description: uri of new object
          description: 'Created'
        '400':
          $ref: "#/components/responses/BadRequest"
        '409':
          $ref: "#/components/responses/Conflict"
        '418':
          $ref: "#/components/responses/ImaTeapot"
        '500':
          $ref: "#/components/responses/InternalError"